# docs/api/company-lists-openapi.yaml
openapi: 3.1.0
info:
  title: Albaly OS - Company Lists API
  version: 1.0.0
  description: >
    Endpoints for user-defined Company Lists and company browsing.
    Matches DB: common_company_lists, common_company_registrations, common_company_contacts, ref_tags, common_company_tags.
servers:
  - url: https://api.albaly.local/v1
security:
  - bearerAuth: []
paths:
  /company-lists:
    get:
      summary: List accessible company lists
      operationId: listCompanyLists
      parameters:
        - in: query
          name: scope
          schema: { type: string, enum: [mine, shared, org] }
          description: Filter by access scope
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 25 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCompanyLists'
    post:
      summary: Create a company list
      operationId: createCompanyList
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompanyListCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CompanyList' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
        traceId: { type: string }
    CompanyListCreate:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        visibility: { type: string, enum: [private, org, public], default: private }
        isShared: { type: boolean, default: false }
...
    CompanyList:
      type: object
      required: [id, name, ownerUserId, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        ownerUserId: { type: string, format: uuid }
        visibility: { type: string, enum: [private, org, public] }
        isShared: { type: boolean }
        itemCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PaginatedCompanyLists:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/CompanyList' }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
...